refactor: 리렌더링 관련 이슈로 인한 리팩토링
- 리렌더링이 필요하지 않은 부분까지 발생하는 문제
- 필요한 부분을 컴포넌트화(최소화)하여 리렌더링이 들어가는 코스트 절감
