+++ client/src/components/UI/molecules/index.ts
@@ -16,3 +16,5 @@ export type { TimeTableTabBtnGroupProps } from './TimeTableTabBtnGroup/TimeTable
 export { TimeTableTabMenu } from './TimeTableTabMenu/TimeTableTabMenu';
 export type { TimeTableTabMenuProps } from './TimeTableTabMenu/TimeTableTabMenu';
 export { LectureBoxContainer } from './LectureBoxContainer/LectureBoxContainer';
+export { BasketLectureListBody } from './LectureListBody/BasketLectureListBody';
+export { SearchedLectureListBody } from './LectureListBody/SearchedLectureListBody';
